// liber-web/src/app/modules/openapi/model/author.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Author
 */
export interface Author { 
    id?: number;
    /**
     * Name
     */
    name: string;
    /**
     * Name
     */
    sort: string;
}