// liber-web/src/app/modules/openapi/model/book.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Author } from './author';
import { Publisher } from './publisher';
import { Tag } from './tag';
import { BookData } from './book-data';
import { Serie } from './serie';


/**
 * Book
 */
export interface Book { 
    authors?: Set<Author>;
    /**
     * Base64 encoded cover image
     */
    cover?: string;
    data?: Set<BookData>;
    /**
     * Book description
     */
    description?: string;
    id?: number;
    /**
     * ISO 639 alpha-2/alpha-3 language code.
     */
    language: string;
    publishers?: Set<Publisher>;
    /**
     * Series the book belongs to
     */
    serie?: Serie;
    /**
     * Series index, if the book is part of a series. If not, it should be 1.
     */
    seriesIndex: number;
    tags?: Set<Tag>;
    /**
     * Book title
     */
    title: string;
}