// liber-web/src/app/modules/openapi/model/serie.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Book } from './book';


/**
 * Serie
 */
export interface Serie { 
    books?: Array<Book>;
    id?: number;
    /**
     * Name
     */
    name: string;
}