// liber-web/src/app/modules/openapi/model/tag.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Tag
 */
export interface Tag { 
    id?: number;
    /**
     * Name
     */
    name: string;
}